requirements.txt
black==24.8.0
google-api-python-client==2.144.0
openai==1.44.0
pip-chill==1.0.3
pytest==8.3.2
pyyaml==6.0.2
youtube-transcript-api==0.6.2
black==24.8.0
google-api-python-client==2.144.0
isort==5.13.2
openai==1.44.0
pip-chill==1.0.3
pytest==8.3.2
pyyaml==6.0.2
youtube-transcript-api==0.6.2
